Enhance type-safety in distributions
- Updated README.md to introduce the new type-safe distribution system, highlighting its benefits and key improvements.
- Added a new section demonstrating type-safe usage of Bernoulli, Poisson, and Categorical distributions with natural return types.
- Introduced a new example file `fully_type_safe.rs` showcasing the type-safe distribution system in action.
- Refactored existing examples to utilize type-safe boolean comparisons and improved clarity in probabilistic modeling.
- Enhanced documentation across various modules to reflect changes in distribution handling and type safety.

+++ examples/conjugate_beta_binomial.rs
@@ -11,7 +11,7 @@ fn beta_binomial_model(n: u64, k: u64) -> Model<f64> {
             beta: 2.0,
         },
     )
-    .bind(move |p| observe(addr!("obs"), Binomial { n, p }, k as f64).bind(move |_| pure(p)))
+    .bind(move |p| observe(addr!("obs"), Binomial { n, p }, k).bind(move |_| pure(p)))
 }
 
 #[derive(Parser, Debug)]
